+++ TinyNvidiaUpdateChecker/LocationChooserForm.Designer.cs
@@ -1,0 +1,103 @@
+﻿namespace TinyNvidiaUpdateChecker
+{
+    partial class LocationChooserForm
+    {
+        /// <summary>
+        /// Required designer variable.
+        /// </summary>
+        private System.ComponentModel.IContainer components = null;
+
+        /// <summary>
+        /// Clean up any resources being used.
+        /// </summary>
+        /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing && (components != null))
+            {
+                components.Dispose();
+            }
+            base.Dispose(disposing);
+        }
+
+        #region Windows Form Designer generated code
+
+        /// <summary>
+        /// Required method for Designer support - do not modify
+        /// the contents of this method with the code editor.
+        /// </summary>
+        private void InitializeComponent()
+        {
+            this.locationListView = new System.Windows.Forms.ListView();
+            this.columnLocation = ((System.Windows.Forms.ColumnHeader)(new System.Windows.Forms.ColumnHeader()));
+            this.ConfirmBtn = new System.Windows.Forms.Button();
+            this.label1 = new System.Windows.Forms.Label();
+            this.SuspendLayout();
+            // 
+            // locationListView
+            // 
+            this.locationListView.Columns.AddRange(new System.Windows.Forms.ColumnHeader[] {
+            this.columnLocation});
+            this.locationListView.FullRowSelect = true;
+            this.locationListView.GridLines = true;
+            this.locationListView.HideSelection = false;
+            this.locationListView.Location = new System.Drawing.Point(12, 51);
+            this.locationListView.MultiSelect = false;
+            this.locationListView.Name = "locationListView";
+            this.locationListView.Size = new System.Drawing.Size(217, 238);
+            this.locationListView.TabIndex = 0;
+            this.locationListView.UseCompatibleStateImageBehavior = false;
+            this.locationListView.View = System.Windows.Forms.View.List;
+            // 
+            // columnLocation
+            // 
+            this.columnLocation.Text = "Location";
+            this.columnLocation.Width = 347;
+            // 
+            // ConfirmBtn
+            // 
+            this.ConfirmBtn.Location = new System.Drawing.Point(12, 295);
+            this.ConfirmBtn.Name = "ConfirmBtn";
+            this.ConfirmBtn.Size = new System.Drawing.Size(75, 23);
+            this.ConfirmBtn.TabIndex = 1;
+            this.ConfirmBtn.Text = "Confirm";
+            this.ConfirmBtn.UseVisualStyleBackColor = true;
+            this.ConfirmBtn.Click += new System.EventHandler(this.ConfirmBtn_Click);
+            // 
+            // label1
+            // 
+            this.label1.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+            this.label1.AutoSize = true;
+            this.label1.Location = new System.Drawing.Point(34, 19);
+            this.label1.Name = "label1";
+            this.label1.Size = new System.Drawing.Size(173, 17);
+            this.label1.TabIndex = 2;
+            this.label1.Text = "Choose download location";
+            // 
+            // LocationChooserForm
+            // 
+            this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.ClientSize = new System.Drawing.Size(241, 330);
+            this.Controls.Add(this.label1);
+            this.Controls.Add(this.ConfirmBtn);
+            this.Controls.Add(this.locationListView);
+            this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.None;
+            this.Name = "LocationChooserForm";
+            this.StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;
+            this.Text = "LocationChooserForm";
+            this.Load += new System.EventHandler(this.LocationChooserForm_Load);
+            this.ResumeLayout(false);
+            this.PerformLayout();
+
+        }
+
+        #endregion
+
+        private System.Windows.Forms.ListView locationListView;
+        private System.Windows.Forms.ColumnHeader columnLocation;
+        private System.Windows.Forms.Button ConfirmBtn;
+        private System.Windows.Forms.Label label1;
+    }
+}